+++ NewTek_Dashboard SRR - VS2013/Source/TestServer/TestServer.cpp
@@ -154,7 +154,7 @@ void CommandLineInterface()
 					{
 						if (stricmp(TestName[i],str_1)==0)
 						{
-							Test=i;
+							Test=(int)i;
 							break;
 						}
 					}
